+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The types of changes are:
 
 ### Added
 - Access and erasure support for Aircall [#2589](https://github.com/ethyca/fides/pull/2589)
+- Access and erasure support for Klaviyo [#2501](https://github.com/ethyca/fides/pull/2501)
 
 ### Changed
 - The `cursor` pagination strategy now also searches for data outside of the `data_path` when determining the cursor value [#3068](https://github.com/ethyca/fides/pull/3068)